fastapi/requirements.txt
aioredis==1.3.1
elasticsearch[async]==7.9.1
fastapi==0.61.1
pydantic==1.9.0
uvicorn[standard]==0.12.2
orjson==3.6.7
gunicorn==20.1.0
backoff==1.11.1
grpcio==1.46.0
grpcio-tools==1.46.0